// src/types/slackdump.d.ts
interface SlackMessage {
  client_msg_id: string;
  type: string;
  user: string;
  text: string;
  ts: string;
  edited?: {
    user: string;
    ts: string;
  };
  files?: SlackFile[];
  reactions?: Reaction[];
  replace_original: boolean;
  delete_original: boolean;
  metadata: {
    event_type: string;
    event_payload: unknown | null;
  };
  blocks?: Block[];
  user_team: string;
  source_team: string;
  user_profile: UserProfile;
  reply_users_count: number;
}

interface SlackFile {
  id: string;
  created: number;
  timestamp: number;
  name: string;
  title: string;
  mimetype: string;
  image_exif_rotation: number;
  filetype: string;
  pretty_type: string;
  user: string;
  mode: string;
  editable: boolean;
  is_external: boolean;
  external_type: string;
  size: number;
  url: string;
  url_download: string;
  url_private: string;
  url_private_download: string;
  original_h: number;
  original_w: number;
  thumb_64: string;
  thumb_80: string;
  thumb_160: string;
  thumb_360: string;
  thumb_360_gif: string;
  thumb_360_w: number;
  thumb_360_h: number;
  thumb_480: string;
  thumb_480_w: number;
  thumb_480_h: number;
  thumb_720: string;
  thumb_720_w: number;
  thumb_720_h: number;
  thumb_960: string;
  thumb_960_w: number;
  thumb_960_h: number;
  thumb_1024: string;
  thumb_1024_w: number;
  thumb_1024_h: number;
  permalink: string;
  permalink_public: string;
  edit_link: string;
  preview: string;
  preview_highlight: string;
  lines: number;
  lines_more: number;
  is_public: boolean;
  public_url_shared: boolean;
  channels: string[] | null;
  groups: string[] | null;
  ims: string[] | null;
  initial_comment: unknown;
  comments_count: number;
  num_stars: number;
  is_starred: boolean;
  shares: {
    public: unknown | null;
    private: unknown | null;
  };
}

interface Reaction {
  name: string;
  count: number;
  users: string[];
}

interface Block {
  type: string;
  block_id: string;
  elements: Element[];
}

interface Element {
  type: string;
  elements?: Element[];
  text?: string;
  name?: string;
  skin_tone?: number;
  user_id?: string;
}

interface UserProfile {
  avatar_hash: string;
  image_72: string;
  first_name: string;
  real_name: string;
  display_name: string;
  team: string;
  name: string;
  is_restricted: boolean;
  is_ultra_restricted: boolean;
}